+++ tests/adaptors/test_file.py
@@ -1,5 +1,5 @@
 import pytest
-from qcio import FileSpec
+from qcio import FileInput
 
 from qcop.adapters import FileAdapter
 from qcop.adapters.utils import tmpdir
@@ -7,8 +7,8 @@ from qcop.exceptions import ProgramNotFoundError
 
 
 def test_file_adapter_compute(tmp_path):
-    # Create FileSpec
-    file_inp = FileSpec(cmdline_args=["hello_world.py"])
+    # Create FileInput
+    file_inp = FileInput(cmdline_args=["hello_world.py"])
     file_inp.files["hello_world.py"] = "print('hello world')"
     with tmpdir() as path:
         file_inp.save_files(path)
@@ -17,7 +17,7 @@ def test_file_adapter_compute(tmp_path):
 
 
 def test_file_adapter_raises_exception_if_program_fails():
-    # Create FileSpec
-    file_inp = FileSpec()
+    # Create FileInput
+    file_inp = FileInput()
     with pytest.raises(ProgramNotFoundError):
         FileAdapter("does_not_exist").compute_data(file_inp)