+++ tests/integration_tests/test_geometric_integration.py
@@ -1,5 +1,5 @@
 import pytest
-from qcio import CalcType, CompositeCalcSpec, OptimizationResults, Results
+from qcio import CalcType, DualProgramInput, OptimizationResults, Results
 
 from qcop.adapters import GeometricAdapter
 from tests.conftest import skipif_program_not_available
@@ -7,16 +7,16 @@ from tests.conftest import skipif_program_not_available
 
 @pytest.mark.integration
 @skipif_program_not_available("terachem")
-def test_full_optimization(dual_prog_inp):
-    prog_inp = dual_prog_inp(CalcType.optimization)
-    prog_inp_dict = prog_inp.model_dump()
-    prog_inp_dict["subprogram"] = "terachem"
-    prog_inp = CompositeCalcSpec(**prog_inp_dict)
+def test_full_optimization(dual_prog_input_factory):
+    prog_input = dual_prog_input_factory(CalcType.optimization)
+    prog_input_dict = prog_input.model_dump()
+    prog_input_dict["subprogram"] = "terachem"
+    prog_input = DualProgramInput(**prog_input_dict)
 
     adapter = GeometricAdapter()
-    output = adapter.compute(prog_inp, propagate_wfn=True)
+    output = adapter.compute(prog_input, propagate_wfn=True)
     assert isinstance(output, Results)
-    assert isinstance(output.input_data, CompositeCalcSpec)
+    assert isinstance(output.input_data, DualProgramInput)
     assert isinstance(output.data, OptimizationResults)
     # Ensure wavefunction was propagated
     assert (
@@ -28,22 +28,22 @@ def test_full_optimization(dual_prog_inp):
 
 @pytest.mark.integration
 @skipif_program_not_available("terachem")
-def test_full_transition_state(dual_prog_inp, water):
+def test_full_transition_state(dual_prog_input_factory, water):
     """This test lacks any test of correctness, but it does ensure that the
     transition state search does not fail.
     """
     # Must use water or else the transition state search will fail
-    prog_inp = dual_prog_inp(CalcType.transition_state)
-    prog_inp_dict = prog_inp.model_dump()
-    prog_inp_dict["subprogram"] = "terachem"
-    prog_inp_dict["structure"] = water
-    prog_inp = CompositeCalcSpec(**prog_inp_dict)
+    prog_input = dual_prog_input_factory(CalcType.transition_state)
+    prog_input_dict = prog_input.model_dump()
+    prog_input_dict["subprogram"] = "terachem"
+    prog_input_dict["structure"] = water
+    prog_input = DualProgramInput(**prog_input_dict)
 
     adapter = GeometricAdapter()
     # Ensure output was produced
-    output = adapter.compute(prog_inp, propagate_wfn=True)
+    output = adapter.compute(prog_input, propagate_wfn=True)
     assert isinstance(output, Results)
-    assert isinstance(output.input_data, CompositeCalcSpec)
+    assert isinstance(output.input_data, DualProgramInput)
     assert isinstance(output.data, OptimizationResults)
     # Ensure wavefunction was propagated
     assert (